app/requirements.txt
click==8.1.8

crewai==0.130.0
crewai-tools==0.47.1

fastapi==0.115.9
uvicorn==0.29.0
python-multipart==0.0.9

python-dotenv==1.0.1

huggingface-hub==0.23.4

pypdf==5.1.0

numpy==1.26.4
pandas==2.2.2

protobuf==4.25.3

litellm>=1.35.0